Fix: Applied MR feedback
