+++ chat/src/main/java/com/github/twitch4j/chat/events/channel/ChannelMessageEvent.java
@@ -1,7 +1,6 @@
 package com.github.twitch4j.chat.events.channel;
 
-import com.github.twitch4j.chat.events.AbstractChannelEvent;
-import com.github.twitch4j.chat.flag.AutoModFlag;
+import com.github.twitch4j.chat.events.AbstractChannelMessageEvent;
 import com.github.twitch4j.chat.util.ChatCrowdChant;
 import com.github.twitch4j.common.annotation.Unofficial;
 import com.github.twitch4j.common.enums.CommandPermission;
@@ -13,7 +12,6 @@ import lombok.Getter;
 import lombok.Value;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
@@ -22,37 +20,7 @@ import java.util.Set;
  */
 @Value
 @EqualsAndHashCode(callSuper = false)
-public class ChannelMessageEvent extends AbstractChannelEvent implements ReplyableEvent {
-
-    /**
-     * RAW Message Event
-     */
-    private final IRCMessageEvent messageEvent;
-
-    /**
-     * User
-     */
-    private EventUser user;
-
-    /**
-     * Message
-     */
-    private String message;
-
-    /**
-     * Permissions of the user
-     */
-    private Set<CommandPermission> permissions;
-
-    /**
-     * The exact number of months the user has been a subscriber, or zero if not subscribed
-     */
-    private int subscriberMonths;
-
-    /**
-     * The tier at which the user is subscribed (prime is treated as 1), or zero if not subscribed
-     */
-    private int subscriptionTier;
+public class ChannelMessageEvent extends AbstractChannelMessageEvent {
 
     /**
      * Nonce
@@ -85,14 +53,13 @@ public class ChannelMessageEvent extends AbstractChannelEvent implements Replyab
      * @param message      The plain text of the message.
      * @param permissions  The permissions of the triggering user.
      */
-    public ChannelMessageEvent(EventChannel channel, IRCMessageEvent messageEvent, EventUser user, String message, Set<CommandPermission> permissions) {
-        super(channel);
-        this.messageEvent = messageEvent;
-        this.user = user;
-        this.message = message;
-        this.permissions = permissions;
-        this.subscriberMonths = messageEvent.getSubscriberMonths().orElse(0);
-        this.subscriptionTier = messageEvent.getSubscriptionTier().orElse(0);
+    public ChannelMessageEvent(
+        EventChannel channel,
+        IRCMessageEvent messageEvent,
+        EventUser user, String message,
+        Set<CommandPermission> permissions
+    ) {
+        super(channel, messageEvent, user, message, permissions);
     }
 
     /**
@@ -138,13 +105,4 @@ public class ChannelMessageEvent extends AbstractChannelEvent implements Replyab
     public boolean isUserIntroduction() {
         return "user-intro".equals(getMessageEvent().getTags().get("msg-id"));
     }
-
-    /**
-     * @return the regions of the message that were flagged by AutoMod.
-     */
-    @Unofficial
-    public List<AutoModFlag> getFlags() {
-        return this.messageEvent.getFlags();
-    }
-
 }
