+++ kotlin/src/main/java/com/github/twitch4j/kotlin/main/EventManagerKtx.kt
@@ -1,4 +1,4 @@
-package com.github.twitch4j.ktx.main
+package com.github.twitch4j.kotlin.main
 
 import com.github.philippheuer.events4j.core.EventManager
 import kotlinx.coroutines.channels.awaitClose
@@ -12,6 +12,7 @@ import kotlinx.coroutines.flow.callbackFlow
  *
  * @return  A Flow object that encapsulates the eventListener and handles the subscription
  */
+@Deprecated("Remove when kotlin module was merged in events4j")
 inline fun <reified T> EventManager.flowOn(): Flow<T> = flowOn(T::class.java)
 
 /**
@@ -22,6 +23,7 @@ inline fun <reified T> EventManager.flowOn(): Flow<T> = flowOn(T::class.java)
  * @param klass The class to receive events for
  * @return      A Flow object that encapsulates the eventListener and handles the subscription
  */
+@Deprecated("Remove when kotlin module was merged in events4j")
 fun <T> EventManager.flowOn(
 	klass: Class<T>
 ): Flow<T> = callbackFlow {
