+++ settings.gradle.kts
@@ -12,7 +12,7 @@ include(
 	":pubsub",
 	":graphql",
 	":twitch4j",
-	":twitch4j-ktx"
+	":kotlin"
 )
 
 project(":common").name = "twitch4j-common"
@@ -26,4 +26,4 @@ project(":rest-tmi").name = "twitch4j-messaginginterface"
 project(":pubsub").name = "twitch4j-pubsub"
 project(":graphql").name = "twitch4j-graphql"
 project(":twitch4j").name = "twitch4j"
-project(":twitch4j-ktx").name = "twitch4j-ktx"
+project(":kotlin").name = "twitch4j-kotlin"
